requirements.txt
dj-database-url==0.3.0
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.4.5
MySQL-python==1.2.5
mysqlclient==1.3.7
pbr==1.9.1
psycopg2==2.6.1
requests==2.9.1
six==1.10.0
stevedore==1.13.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
whitenoise==2.0.6

